pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "issue-reporter"
dependencies=["jira", "pyyaml", "jinja2"]
version = "0.0.1"
authors = [
  { name="Megan Wilhite", email="megan.wilhite@gmail.com" },
]
description = "A tool to return issue reports"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/Ch3LL/issue-reporter"
Issues = "https://github.com/Ch3LL/issue-reporter/issues"
